+++ unpythonic/syntax/tests/test_letdo.py
@@ -17,22 +17,47 @@ from ...seq import begin
 x = "the global x"  # for lexical scoping tests
 
 def runtests():
-    with testset("do (imperative code in an expression)"):
+    with testset("do (imperative code in an expression) (new env-assignment syntax 0.15.3+)"):
         # Macro wrapper for unpythonic.seq.do (imperative code in expression position)
-        #  - Declare and initialize a local variable with ``local[var << value]``.
+        #  - Declare and initialize a local variable with ``local[var := value]``.
         #    Is in scope from the next expression onward, for the (lexical) remainder
         #    of the do.
-        #  - Assignment is ``var << value``. Valid from any level inside the ``do``
+        #  - Assignment is ``var := value``. Valid from any level inside the ``do``
         #    (including nested ``let`` constructs and similar).
         #  - No need for ``lambda e: ...`` wrappers. Inserted automatically,
         #    so the lines are only evaluated as the underlying seq.do() runs.
+        d1 = do[local[x := 17],
+                print(x),
+                x := 23,
+                x]
+        test[d1 == 23]
+
+        # Since we repurposed an existing assignment operator, let's check we didn't accidentally assign to the function scope.
+        test_raises[NameError, x, "only the `do[]` should have an `x` here"]
+
+        # v0.14.0: do[] now supports deleting previously defined local names with delete[]
+        a = 5
+        d = do[local[a := 17],  # noqa: F841, yes, d is unused.
+               test[a == 17],
+               delete[a],
+               test[a == 5],  # lexical scoping
+               True]
+
+        test_raises[KeyError, do[delete[a], ], "should have complained about deleting nonexistent local 'a'"]
+
+        # do0[]: like do[], but return the value of the **first** expression
+        d2 = do0[local[y := 5],  # noqa: F821, `local` defines the name on the LHS of the `<<`.
+                 print("hi there, y =", y),  # noqa: F821
+                 42]  # evaluated but not used
+        test[d2 == 5]
+
+    with testset("do (imperative code in an expression) (previous modern env-assignment syntax)"):
         d1 = do[local[x << 17],
                 print(x),
                 x << 23,
-                x]  # do[] returns the value of the last expression
+                x]  # do[] returns the value of the last expression  # noqa: F823, it's the `x` from `do[]`, not from the enclosing scope.
         test[d1 == 23]
 
-        # v0.14.0: do[] now supports deleting previously defined local names with delete[]
         a = 5
         d = do[local[a << 17],  # noqa: F841, yes, d is unused.
                test[a == 17],
@@ -42,14 +67,41 @@ def runtests():
 
         test_raises[KeyError, do[delete[a], ], "should have complained about deleting nonexistent local 'a'"]
 
-        # do0[]: like do[], but return the value of the **first** expression
         d2 = do0[local[y << 5],  # noqa: F821, `local` defines the name on the LHS of the `<<`.
                  print("hi there, y =", y),  # noqa: F821
                  42]  # evaluated but not used
         test[d2 == 5]
 
     # Let macros. Lexical scoping supported.
-    with testset("let, letseq, letrec basic usage"):
+    with testset("let, letseq, letrec basic usage (new env-assignment syntax 0.15.3+)"):
+        # parallel binding, i.e. bindings don't see each other
+        test[let[x := 17,
+                 y := 23][  # noqa: F821, `let` defines `y` here.
+                     (x, y)] == (17, 23)]  # noqa: F821
+
+        # sequential binding, i.e. Scheme/Racket let*
+        test[letseq[x := 1,
+                    y := x + 1][  # noqa: F821
+                        (x, y)] == (1, 2)]  # noqa: F821
+
+        test[letseq[x := 1,
+                    x := x + 1][  # in a letseq, rebinding the same name is ok
+                        x] == 2]
+
+        # letrec sugars unpythonic.lispylet.letrec, removing the need for quotes on LHS
+        # and "lambda e: ..." wrappers on RHS (these are inserted by the macro):
+        test[letrec[evenp := (lambda x: (x == 0) or oddp(x - 1)),  # noqa: F821, `letrec` defines `evenp` here.
+                    oddp := (lambda x: (x != 0) and evenp(x - 1))][  # noqa: F821
+                        evenp(42)] is True]  # noqa: F821
+
+        # nested letrecs work, too - each environment is internally named by a gensym
+        # so that outer ones "show through":
+        test[letrec[z := 9000][  # noqa: F821
+                 letrec[evenp := (lambda x: (x == 0) or oddp(x - 1)),  # noqa: F821
+                        oddp := (lambda x: (x != 0) and evenp(x - 1))][  # noqa: F821
+                            (evenp(42), z)]] == (True, 9000)]  # noqa: F821
+
+    with testset("let, letseq, letrec basic usage (previous modern env-assignment syntax)"):
         # parallel binding, i.e. bindings don't see each other
         test[let[x << 17,
                  y << 23][  # noqa: F821, `let` defines `y` here.
@@ -98,6 +150,32 @@ def runtests():
                     "should not be able to rebind the same name in the same let"]
 
     # implicit do: an extra set of brackets denotes a multi-expr body
+    with testset("implicit do (extra bracket syntax for multi-expr let body) (new env-assignment syntax v0.15.3+)"):
+        a = let[x := 1,
+                y := 2][[  # noqa: F821
+                    y := 1337,  # noqa: F821
+                    (x, y)]]  # noqa: F821
+        test[a == (1, 1337)]
+
+        # only the outermost extra brackets denote a multi-expr body
+        a = let[(x, 1),
+                (y, 2)][[  # noqa: F821
+                    [1, 2]]]
+        test[a == [1, 2]]
+
+        # implicit do works also in letseq, letrec
+        a = letseq[x := 1,
+                   y := x + 1][[  # noqa: F821
+                       x := 1337,
+                       (x, y)]]  # noqa: F821
+        test[a == (1337, 2)]
+
+        a = letrec[x := 1,
+                   y := x + 1][[  # noqa: F821
+                       x := 1337,
+                       (x, y)]]  # noqa: F821
+        test[a == (1337, 2)]
+
     with testset("implicit do (extra bracket syntax for multi-expr let body)"):
         a = let[x << 1,
                 y << 2][[  # noqa: F821